integrations/adk/requirements.txt
google-adk>=1.0
google-genai>=1.0
mubit-sdk>=0.1.0
